pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "etude"
version = "0.1.0"
description = "A controllable piano cover generation framework"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10,<3.13"
authors = [
    {name = "Xiugapurin"}
]
keywords = ["music", "piano-cover-generation", "deep-learning", "transformer"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Analysis",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core ML frameworks (torch/torchaudio will be installed from PyTorch CUDA index)
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "torchcodec>=0.4.0",
    "transformers>=4.30.0",

    # Audio processing
    "librosa>=0.10.0",
    "synctoolbox>=1.3.0",
    "madmom @ git+https://github.com/CPJKU/madmom.git@27f032e8947204902c675e5e341a3faf5dc86dae",

    # MIDI processing
    "mido>=1.3.0",
    "pretty_midi>=0.2.10",

    # Utilities
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "tqdm>=4.65.0",
    "PyYAML>=6.0",
    "yt-dlp>=2023.0.0",
    "pydantic>=2.0.0",

    # Jupyter support
    "ipykernel>=6.0.0",
    "ipywidgets>=8.0.0",
    "matplotlib>=3.7.0",
]

[project.optional-dependencies]
# Alternative audio separation backend using Demucs (replaces Spleeter)
# Install with: uv sync --extra demucs
# Note: Spleeter is the default backend and requires a separate conda environment (py38_spleeter)
demucs = [
    "demucs>=4.0.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://xiugapurin.github.io/Etude/"
Repository = "https://github.com/Xiugapurin/Etude"
Paper = "https://arxiv.org/abs/2509.16522"

[project.scripts]
etude-infer = "etude.cli.infer:main"
etude-prepare = "etude.cli.prepare:main"
etude-train = "etude.cli.train:main"
etude-evaluate = "etude.cli.evaluate:main"
etude-separate = "etude.scripts.run_separation:main"

# =============================================================================
# UV Configuration
# =============================================================================

[tool.uv]
# Use Chinese mirrors for faster download
# PyPI mirror: Tsinghua
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"

# Allow uv to find packages from multiple indexes (PyPI + PyTorch)
index-strategy = "unsafe-best-match"

[tool.uv.sources]
# Pin torch packages to use CUDA index (NJU mirror - China)
torch = { index = "pytorch-cu124" }
torchaudio = { index = "pytorch-cu124" }

[[tool.uv.index]]
name = "pytorch-cu124"
# NJU (Nanjing University) PyTorch wheels mirror (CUDA 12.4)
url = "https://mirrors.nju.edu.cn/pytorch/whl/cu124"
explicit = true

# =============================================================================
# Hatch Build Configuration
# =============================================================================

[tool.hatch.metadata]
# Allow git+ direct references for madmom
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["etude"]

# Include package data files (configs, etc.)
[tool.hatch.build.targets.wheel.sources]
"etude" = "etude"

# Ensure non-Python files are included
[tool.hatch.build]
include = [
    "etude/**/*.py",
    "etude/**/*.yaml",
]

# =============================================================================
# Tool Configuration
# =============================================================================

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
